[PATCH] sched: Make sched_feat(x) boolean for !CONFIG_SCHED_DEBUG
If scheduler debugging is disabled sched_feat() resolves to a constant
value that can be non-boolean. clang raises a warning when it detects
that a non-boolean constant is used as operand in logical expressions:

kernel/sched/core.c:208:26: warning: use of logical '&&' with constant
  operand [-Wconstant-logical-operand]
    if ((irq_delta + steal) && sched_feat(NONTASK_CAPACITY))
                            ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Convert the original expression to boolean by means of a double not
operation.

Reported-by: Philipp Klocke <Phil_K97@gmx.de>
Suggested-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>