genirq: Make sure the affinity to be set is not empty
If the default_affinity is managed/initialized by a system and
all the cpus present in irq_default_affinity are hotplugged out
we may end up passing a cpumask of 0 to irq_do_set_affinity().

Fix this by falling back to cpu_online_mask in case the
calculated affinity becomes zero.

Change-Id: I11f6e6dd7e44865330296f7c48aa2d0594d31610
Signed-off-by: Srinivas Ramana <sramana@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>